(PC-14451)[API] refactor: move sending_limit to fraud/phone_validation
